+++ features/cucumber.mjs
@@ -1,0 +1,9 @@
+export default {
+  default: {
+    paths: ['features/**/*.feature'],
+    import: ['features/**/*.js'],
+    format: ['progress'],
+    publishQuiet: true,
+    parallel: 1,
+  },
+}